pyproject.toml
[tool.poetry]
name = "hw-aiohttp"
version = "0.1.0"
description = ""
authors = ["Maksim Molianov <mm_89615@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = {extras = ["speedups"], version = "^3.11.10"}
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.1"
pydantic = "^2.10.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
